Revert "update for windows error C2371"
This reverts commit c05b5912da06752f874e25b9e87e380850d02991.

+++ core/utils/instance_handler.h
@@ -30,7 +30,7 @@ class InstanceHandler {
 
   void Create(int32_t id, std::unique_ptr<type> instance);
 
-  int32_t Count();
+  size_t Count();
 
   void Dispose(int32_t id);
 
@@ -56,7 +56,7 @@ void InstanceHandler<type>::Create(int32_t id, std::unique_ptr<type> instance) {
 }
 
 template <typename type>
-int32_t InstanceHandler<type>::Count() {
+size_t InstanceHandler<type>::Count() {
   std::lock_guard<std::mutex> lock(mutex_);
   return instances.size();
 }
